added option for RDBI requests

+++ sw/firmware/components/can_driver/include/obd2_dtb.hpp
@@ -19,48 +19,48 @@ class OBD2DTB
 {
 public:
     void initDef();
-    bool isSup(uint8_t pid) const;
-    bool pidExists(uint8_t pid) const;
+    bool isSup(uint16_t pid) const;
+    bool pidExists(uint16_t pid) const;
     // PID_DEF Getters
-    uint8_t getmode(uint8_t pid) const;
-    const char *getName(uint8_t pid) const;
-    const char *getUnit(uint8_t pid) const;
-    const char *getDescription(uint8_t pid) const;
-    float getMinValue(uint8_t pid) const;
-    float getMaxValue(uint8_t pid) const;
-    uint8_t getPriority(uint8_t pid) const;
-    uint16_t getUpdateInterval(uint8_t pid) const;
-    uint32_t getColor(uint8_t pid) const;
-    const char *getIcon(uint8_t pid) const;
-    const char *getFormula(uint8_t pid) const;
-    PIDDefinition *getDef(uint8_t pid) const;
-    esp_err_t getDef(uint8_t pid, const PIDDefinition *&outDef) const;
+    uint8_t getmode(uint16_t pid) const;
+    const char *getName(uint16_t pid) const;
+    const char *getUnit(uint16_t pid) const;
+    const char *getDescription(uint16_t pid) const;
+    float getMinValue(uint16_t pid) const;
+    float getMaxValue(uint16_t pid) const;
+    uint8_t getPriority(uint16_t pid) const;
+    uint16_t getUpdateInterval(uint16_t pid) const;
+    uint32_t getColor(uint16_t pid) const;
+    const char *getIcon(uint16_t pid) const;
+    const char *getFormula(uint16_t pid) const;
+    PIDDefinition *getDef(uint16_t pid) const;
+    esp_err_t getDef(uint16_t pid, const PIDDefinition *&outDef) const;
 
     // Special Getters
     std::string getVIN() const;
     std::vector<std::string> getDTC(uint8_t mode);
 
     // PID Data Getters
-    float getValue(uint8_t pid, uint32_t timeout_ms = 500) const;
-    uint32_t getLastUpdated(uint8_t pid) const;
-    uint32_t getId(uint8_t pid) const;
-    esp_err_t getRawData(uint8_t pid, uint8_t *outData) const;
-    bool isValid(uint8_t pid) const;
-    esp_err_t getData(uint8_t pid, PIDData_t &pd) const;
+    float getValue(uint16_t pid, uint32_t timeout_ms = 500) const;
+    uint32_t getLastUpdated(uint16_t pid) const;
+    uint32_t getId(uint16_t pid) const;
+    esp_err_t getRawData(uint16_t pid, uint8_t *outData) const;
+    bool isValid(uint16_t pid) const;
+    esp_err_t getData(uint16_t pid, PIDData_t &pd) const;
 
     // PID Data Setters
-    esp_err_t setUpdateInterval(uint8_t pid, UpdateRate interval_ms);
-    esp_err_t setValid(uint8_t pid, bool valid);
-    esp_err_t setIsSupported(uint8_t pid, bool supported);
-    esp_err_t setLastUpdated(uint8_t pid, uint32_t lastUpdated);
-    esp_err_t setId(uint8_t pid, uint32_t id);
+    esp_err_t setUpdateInterval(uint16_t pid, UpdateRate interval_ms);
+    esp_err_t setValid(uint16_t pid, bool valid);
+    esp_err_t setIsSupported(uint16_t pid, bool supported);
+    esp_err_t setLastUpdated(uint16_t pid, uint32_t lastUpdated);
+    esp_err_t setId(uint16_t pid, uint32_t id);
     esp_err_t setDTC(uint16_t rawDTC, uint8_t mode);
     esp_err_t clearDTC(uint8_t mode);
     std::string decodeDTC(uint16_t rawDTC);
-    std::vector<uint8_t> getPIDs() const;
+    std::vector<uint16_t> getPIDs() const;
     esp_err_t updateData(const CanDriver::CanFrame &frame);
 
-    using DataUpdateCallback = std::function<void(uint8_t pid)>;
+    using DataUpdateCallback = std::function<void(uint16_t pid)>;
     void subscribe(DataUpdateCallback cb);
     std::vector<DataUpdateCallback> subscribers_;
 
@@ -84,13 +84,13 @@ public:
         return PID_DEF.size();
     }
 
-    void addPID(uint8_t mode, uint8_t pid, std::string name, std::string unit,
+    void addPID(uint8_t mode, uint16_t pid, std::string name, std::string unit,
                 std::string desc, std::string formula, float minV, float maxV,
                 uint8_t priority, UpdateRate interval, uint32_t color, std::string icon);
 
     // PID Definitions and Data Storage
-    std::map<uint8_t, std::unique_ptr<PIDDefinition>> PID_DEF;
-    std::map<uint8_t, PIDData_t> pidData;
+    std::map<uint16_t, std::unique_ptr<PIDDefinition>> PID_DEF;
+    std::map<uint16_t, PIDData_t> pidData;
     VINData_t vinData;
     DTCData_t dtcData;
