+++ sw/firmware/components/can_driver/include/pid_def.hpp
@@ -12,7 +12,7 @@ class PIDDefinition
 {
 private:
     uint8_t mode_;
-    uint8_t pid_;
+    uint16_t pid_;
     std::string name_;
     std::string unit_;
     std::string description_;
@@ -29,7 +29,7 @@ private:
     static inline SemaphoreHandle_t eval_mutex = xSemaphoreCreateMutex();
 
 public:
-    PIDDefinition(uint8_t mode, uint8_t pid, std::string name, std::string unit,
+    PIDDefinition(uint8_t mode, uint16_t pid, std::string name, std::string unit,
                   std::string desc, std::string formula, float minV, float maxV,
                   uint8_t priority, UpdateRate interval, uint32_t color, std::string icon)
         : mode_(mode), pid_(pid), name_(std::move(name)), unit_(std::move(unit)),
@@ -49,7 +49,7 @@ public:
 
     // Getters
     uint8_t mode() const { return mode_; }
-    uint8_t pid() const { return pid_; }
+    uint16_t pid() const { return pid_; }
     const std::string &name() const { return name_; }
     const std::string &unit() const { return unit_; }
     const std::string &description() const { return description_; }
