+++ sw/firmware/main/include/middleware.hpp
@@ -31,6 +31,6 @@ cJSON *m_dtc_request(int mode);
 cJSON *m_clear_dtc_request();
 
 cJSON *get_single_pid_def_json(const PIDDef_t &pi);
-cJSON *get_single_pid_data_json(uint8_t pid, const PIDData_t &pd);
+cJSON *get_single_pid_data_json(uint16_t pid, const PIDData_t &pd);
 
-esp_err_t get_pid_stream_packet(uint8_t pid, uint8_t *out_packet);
+esp_err_t get_pid_stream_packet(uint16_t pid, uint8_t *out_packet);
